treewide: move code into app subdirectory
This is in preparation of refactoring the code to follow the FastAPI
bigger applications documentation.
