+++ app/internal/command_endpoints/ha_rest.py
