+++ app/internal/command_endpoints/ha_ws.py
