+++ app/internal/command_endpoints/openhab.py
