+++ entrypoint.sh
@@ -10,6 +10,6 @@ set +a
 
 python /app/misc/migrate_devices.py
 
-uvicorn api:app --host 0.0.0.0 --port 8502 --log-config uvicorn-log-config.json \
+uvicorn app.main:app --host 0.0.0.0 --port 8502 --log-config uvicorn-log-config.json \
     --log-level "$LOG_LEVEL" --loop uvloop --timeout-graceful-shutdown 5 \
     --no-server-header
